Restyled code
