+++ include/Internal/error_msg_helpers.h
@@ -20,7 +20,7 @@ namespace Internal {
  */
 inline std::string divide_by_zero_error_msg()
 {
-  return std::string("Potential divide-by-zero.");
+    return std::string("Potential divide-by-zero.");
 }
 
 /**
@@ -35,13 +35,14 @@ inline std::string divide_by_zero_error_msg()
  * @see MathUtils::Vector
  * @see Mathutils::Quaternion
  */
-inline std::string invalid_init_list_length_error_msg(const std::size_t input_len, const std::size_t expected_len)
+inline std::string invalid_init_list_length_error_msg(const std::size_t input_len,
+    const std::size_t expected_len)
 {
-  return  std::string("Expected ") +
-          std::to_string(expected_len) +
-          std::string(", got ") +
-          std::to_string(input_len) +
-          std::string(".");
+    return std::string("Expected ") +
+        std::to_string(expected_len) +
+        std::string(", got ") +
+        std::to_string(input_len) +
+        std::string(".");
 }
 
 /**
@@ -58,23 +59,22 @@ inline std::string invalid_init_list_length_error_msg(const std::size_t input_le
  * @see MathUtils::Matrix
  */
 inline std::string invalid_init_list_length_error_msg(const std::size_t input_rows,
-  const std::size_t input_cols,
-  const std::size_t expected_rows,
-  const std::size_t expected_cols
-  )
+    const std::size_t input_cols,
+    const std::size_t expected_rows,
+    const std::size_t expected_cols)
 {
-  return  std::string("Expected ") +
-          std::string("(") +
-          std::to_string(expected_rows) +
-          std::string(",") +
-          std::to_string(expected_cols) +
-          std::string(")") +
-          std::string(", got ") +
-          std::string("(") +
-          std::to_string(input_rows) +
-          std::string(",") +
-          std::to_string(input_cols) +
-          std::string(").");
+    return std::string("Expected ") +
+        std::string("(") +
+        std::to_string(expected_rows) +
+        std::string(",") +
+        std::to_string(expected_cols) +
+        std::string(")") +
+        std::string(", got ") +
+        std::string("(") +
+        std::to_string(input_rows) +
+        std::string(",") +
+        std::to_string(input_cols) +
+        std::string(").");
 }
 
 /**
@@ -89,13 +89,14 @@ inline std::string invalid_init_list_length_error_msg(const std::size_t input_ro
  * @see MathUtils::Vector
  * @see Mathutils::Quaternion
  */
-inline std::string invalid_index_error_msg(const std::size_t input_index, const std::size_t max_length)
+inline std::string invalid_index_error_msg(const std::size_t input_index,
+    const std::size_t max_length)
 {
-  return  std::string("Index ") +
-          std::to_string(input_index) +
-          std::string(" is invalid for ") +
-          std::to_string(max_length) +
-          std::string(" size object.");
+    return std::string("Index ") +
+        std::to_string(input_index) +
+        std::string(" is invalid for ") +
+        std::to_string(max_length) +
+        std::string(" size object.");
 }
 
 /**
@@ -111,27 +112,28 @@ inline std::string invalid_index_error_msg(const std::size_t input_index, const 
  *
  * @see MathUtils::Matrix
  */
-inline std::string invalid_index_error_msg(const std::size_t input_row, const std::size_t input_col,
-  const std::size_t max_rows,
-  const std::size_t max_cols)
+inline std::string invalid_index_error_msg(const std::size_t input_row,
+    const std::size_t input_col,
+    const std::size_t max_rows,
+    const std::size_t max_cols)
 {
-  return  std::string("Index ") +
-          std::string("(") +
-          std::to_string(input_row) +
-          std::string(",") +
-          std::to_string(input_col) +
-          std::string(")") +
-          std::string(" is invalid for ") +
-          std::string("(") +
-          std::to_string(max_rows) +
-          std::string(",") +
-          std::to_string(max_cols) +
-          std::string(")") +
-          std::string(" size object.");
+    return std::string("Index ") +
+        std::string("(") +
+        std::to_string(input_row) +
+        std::string(",") +
+        std::to_string(input_col) +
+        std::string(")") +
+        std::string(" is invalid for ") +
+        std::string("(") +
+        std::to_string(max_rows) +
+        std::string(",") +
+        std::to_string(max_cols) +
+        std::string(")") +
+        std::string(" size object.");
 }
 
-}  // namespace Internal
-}  // namespace MathUtils
+}    // namespace Internal
+}    // namespace MathUtils
 
 
-#endif  // MATHUTILS_INTERNAL_ERROR_MESSAGE_HELPERS_H_
+#endif    // MATHUTILS_INTERNAL_ERROR_MESSAGE_HELPERS_H_
