+++ include/acos_safe.h
@@ -7,13 +7,13 @@
 #ifndef MATHUTILS_ACOS_SAFE_H_
 #define MATHUTILS_ACOS_SAFE_H_
 
-
 namespace MathUtils {
 
 /**
  * @brief Compute arcosine with bounds checks. Limits output to [pi, 0].
  *
- * @details If the input is greater than 1, 0 is returned. If the input is less than -1, pi is returned.
+ * @details If the input is greater than 1, 0 is returned. If the input is less than -1, pi
+ * is returned.
  *
  * @param val Value to take the arccosine of, in [rad].
  * @return Arccosine of `val`.
