+++ include/asin_safe.h
@@ -12,7 +12,8 @@ namespace MathUtils {
 /**
  * @brief Compute arcsine with bounds checks. Limits output to [-pi/2, pi/2].
  *
- * @details If the input is greater than 1, pi/2 is returned. If the input is less than -1, -pi/2 is returned.
+ * @details If the input is greater than 1, pi/2 is returned. If the input is less than -1, -pi/2
+ * is returned.
  *
  * @param val Value to take the arcsine of, in [rad].
  * @return Arcsine of `val`.
