+++ src/Geodesy/lla_to_ecef.cpp
@@ -4,8 +4,9 @@
  * @date 2023-03-29
  */
 
-#include "constants.h"
 #include "Geodesy/lla_to_ecef.h"
+
+#include "constants.h"
 
 #include <cmath>
 
@@ -13,21 +14,21 @@ namespace MathUtils {
 
 Vector<3> lla_to_ecef(const GeoCoord& lla)
 {
-  // pre-compute terms
-  const double sin_lat_rad = std::sin(lla.latitude());
-  const double cos_lat_rad = std::cos(lla.latitude());
+    // pre-compute terms
+    const double sin_lat_rad = std::sin(lla.latitude());
+    const double cos_lat_rad = std::cos(lla.latitude());
 
-  // "radius of curvature in the meridian"
-  const double c_term = Constants::WGS84_A_M /
-    std::sqrt(1.0 - (Constants::WGS84_ECC2 * sin_lat_rad * sin_lat_rad));
+    // "radius of curvature in the meridian"
+    const double c_term = Constants::WGS84_A_M /
+        std::sqrt(1.0 - (Constants::WGS84_ECC2 * sin_lat_rad * sin_lat_rad));
 
-  const double s_term = c_term * (1.0 - Constants::WGS84_ECC2);
+    const double s_term = c_term * (1.0 - Constants::WGS84_ECC2);
 
-  return Vector<3> {
-    (c_term + lla.altitude()) * cos_lat_rad * std::cos(lla.longitude()),
-    (c_term + lla.altitude()) * cos_lat_rad * std::sin(lla.longitude()),
-    (s_term + lla.altitude()) * sin_lat_rad
-  };
+    return Vector<3> {
+        (c_term + lla.altitude()) * cos_lat_rad * std::cos(lla.longitude()),
+        (c_term + lla.altitude()) * cos_lat_rad * std::sin(lla.longitude()),
+        (s_term + lla.altitude()) * sin_lat_rad
+    };
 }
 
 }  // namespace MathUtils
