+++ test/float_equality_test.cpp
@@ -13,7 +13,7 @@ namespace {
 
 const std::string test_reports_file = std::string("TESTRESULTS-float_equality.xml");
 
-// ====================================================================================================================
+// =================================================================================================
 TEST(FloatEquality, FloatFalse)
 {
   float val1 = 123.0f;
@@ -21,7 +21,7 @@ TEST(FloatEquality, FloatFalse)
   EXPECT_FALSE(MathUtils::float_equality(val1, val2));
 }
 
-// ====================================================================================================================
+// =================================================================================================
 TEST(FloatEquality, FloatFalseSmallNumber)
 {
   float val1 = 12.0f;
@@ -29,7 +29,7 @@ TEST(FloatEquality, FloatFalseSmallNumber)
   EXPECT_FALSE(MathUtils::float_equality(val1, val2));
 }
 
-// ====================================================================================================================
+// =================================================================================================
 TEST(FloatEquality, FloatTrue)
 {
   float val1 = -13.0f;
@@ -37,7 +37,7 @@ TEST(FloatEquality, FloatTrue)
   EXPECT_TRUE(MathUtils::float_equality(val1, val2));
 }
 
-// ====================================================================================================================
+// =================================================================================================
 TEST(FloatEquality, FloatTrueSmallNumber)
 {
   float val1 = 456.0f;
@@ -47,7 +47,7 @@ TEST(FloatEquality, FloatTrueSmallNumber)
 
 
 
-// ====================================================================================================================
+// =================================================================================================
 int main(int argc, char** argv)
 {
   ::testing::GTEST_FLAG(output) = std::string("xml:") + test_reports_file;
